+++ Sem.04/Pract.04/18.cpp
@@ -11,7 +11,7 @@ int abs(int n)
         : -n;
 }
 
-int pow(int n, int k)
+long pow(long n, int k)
 {
     int result = 1;
